SDK-642: Create profile folder, move classes
