+++ test/Yoti.Auth.Tests/TestTools/Profile.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Yoti.Auth.Profile;
 
 namespace Yoti.Auth.Tests.TestTools
 {
