+++ test/Yoti.Auth.Tests/UserProfileTests.cs
@@ -2,6 +2,7 @@
 using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Yoti.Auth.Images;
+using Yoti.Auth.Profile;
 
 namespace Yoti.Auth.Tests
 {
